(feat) Major improvement, no longer need extended test case class
testInstance now accepts any function as an argument to the
constructor, along with any tuple (arguments for the previous
function), and then the string (expectedOutput).

The tests now run in the constructor, so it occurs automatically.
To get the pass/fail result, call the method 'passed' to recieve
a boolean. Unfortunately, it seems that the constructor (and
destructor) cannot be used to return the pass/fail result.